modified  db model

+++ BusinessConnect.DataStore.Model/Dbo/AdvertisementFile.cs
@@ -7,7 +7,7 @@ namespace BusinessConnect.DataStore.Model.Dbo
     public class AdvertisementFile:FileInfo
     {
        
-        public int CampaignAdvertisementId { get; set; }
+        public long CampaignAdvertisementId { get; set; }
 
         public virtual CampaignAdvertisement CampaignAdvertisement { get; set; }
